+++ mymod/gamesrc/cgame/cg_consolecmds.c
@@ -614,7 +614,7 @@ qboolean CG_ConsoleCommand(void) {
 
 	cmd = CG_Argv(0);
 
-	for (i = 0; i < sizeof ( commands) / sizeof ( commands[0]); i++) {
+	for (i = 0; i < (int)(sizeof ( commands) / sizeof ( commands[0])); i++) {
 		if (Q_strequal(cmd, commands[i].cmd)) {
 			commands[i].function();
 			return qtrue;
@@ -635,7 +635,7 @@ so it can perform tab completion
 void CG_InitConsoleCommands(void) {
 	int i;
 
-	for (i = 0; i < sizeof ( commands) / sizeof ( commands[0]); i++) {
+	for (i = 0; i < (int)(sizeof ( commands) / sizeof ( commands[0])); i++) {
 		trap_AddCommand(commands[i].cmd);
 	}
